refactor(svelte): merge surface-handlers into surface-state
Closes #983.

The extract-to-handlers seam was a type cycle plus eleven one-line
forwards — the module behind the seam saw more deps than the one in
front. Fold handlers, brush state, and effects into one surface
controller; keep pointer/keyboard/area-brush pure tables as the
narrow seams.
